backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Muhammad Bilal Khan <lofi.audit@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.3"
python-dotenv = "^1.0.1"
transformers = "^4.44.2"
tensorflow = "^2.17.0"
openai = "^1.42.0"
rich = "^13.7.1"
ffmpy = "^0.4.0"
nltk = "^3.9.1"
moviepy = "^1.0.3"
deepface = "^0.0.93"
google-generativeai = "^0.7.2"
supabase = "^2.7.3"
pydub = "^0.25.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
